Add support for creating a default acl file.
If a project wants to follow the core/drivers pattern, create an
acl file for it if it doesn't have one. Additionally, allow for
the specification of an additional acl file to append to the end
of the generated acl file to allow for specialization on top of the
general case.

Change-Id: I4d5b905befdb8d742dd1aaa148cf4ec1b3355b01
Reviewed-on: https://review.openstack.org/23372
Reviewed-by: James E. Blair <corvus@inaugust.com>
Reviewed-by: Clark Boylan <clark.boylan@gmail.com>
Reviewed-by: Jeremy Stanley <fungi@yuggoth.org>
Approved: Monty Taylor <mordred@inaugust.com>
Tested-by: Jenkins