Stop using pypi mirrors to build our mirror.
Pypi mirrors tend to be out of date which is annoying when we need to
updated our mirror to include the latest version of a particular package
(eg one of the OpenStack python clients). The pip -M option won't always
see the latest version when talking to the pypi mirrors. To avoid this
problem don't use pip -M and talk to pypi directly.

Change-Id: I6ae8950eba005c020b7d3aae59661abc039a2b40
Reviewed-on: https://review.openstack.org/23959
Reviewed-by: Monty Taylor <mordred@inaugust.com>
Reviewed-by: Anita Kuno <akuno@lavabit.com>
Approved: Jeremy Stanley <fungi@yuggoth.org>
Reviewed-by: Jeremy Stanley <fungi@yuggoth.org>
Tested-by: Jenkins