Edited the parsed url to work.
There was duplication in the parsed url, I removed the duplication
to get the url to work.

I commented out the swift heading in the ini file so that the
absence-of-swift conditionals were met in the script.

I set the output_mode to multiple since that was the setting that I
got to work for me so I thought having a working ini file was a good
idea.

*jeepyb/cmd/openstackwatch.py
*jeepyb/confif/openstackwatch.ini-sample

Change-Id: I3e8f8170abd33a99989d96f226562b4c625411fb
Reviewed-on: https://review.openstack.org/26699
Reviewed-by: Chmouel Boudjnah <chmouel@enovance.com>
Reviewed-by: Clark Boylan <clark.boylan@gmail.com>
Approved: James E. Blair <corvus@inaugust.com>
Reviewed-by: James E. Blair <corvus@inaugust.com>
Tested-by: Jenkins