Be more agressive trying to install requirements.
When -r requirements.txt is used, either then entire file works
or the entire file doesn't. This means that if there is a hiccup
with any of the files, we lose the entire run instead of just
the one file.

By iterating over the list and installing as many as we can, then
processing the freeze, we get as many of them as we can.

This is especially helpful when combined with the upcoming pbr/jeepyb
gate, where the lack of a requirement in the gate will show up when
the thing that needs it can't install it. If we get everything else,
the error message in that case will be much more clear as to why.

Change-Id: I27606b92a8be2605fa30362aaab51e65a21770bc