Makes Impact bug creation generic, ceases doc spam
This patch moves the bug creation code to its own method, so that it
can be used by other kinds of impacts.

It also changes the notification logic for DocImpact, so that emails
are now only sent if bug creation was not successful.

This patch should have no effect on other kinds of impacts.

patchset3 fixes use of sys and updated docstring

patchset4 rebases to master

Change-Id: I6b64beb6e8da45b166d17c9fad0347aa5370e689
Reviewed-on: https://review.openstack.org/34842
Approved: James E. Blair <corvus@inaugust.com>
Reviewed-by: James E. Blair <corvus@inaugust.com>
Tested-by: Jenkins