cease docimpact emails, create bugs only
We've been running the DocImpact bug generation code for
several weeks now and are happy with the results. We'd now
like to cease the emails and only generate bugs from DocImpact
strings.

This patch updates the logic to never send an email for DocImpact,
and changes the docstring to reflect this.

This change should not affect any other types of impacts.

patchset2 fixes PEP8 issue

Change-Id: I5c46f0d540949bc4a5de1e57fcd68c537a54ca15