Add distribute to the pre-install list
mysql-python is crapping the fantastic bong because of a continued
reliance on distribute. Fine. We'll just make sure we have latest
no-op package in our virtualenv so that epic fail stops killing bunnies.

Also, when we're running the wheel command in our venvs, we kinda need
wheel to be installed in them.

Change-Id: I004ca4efdbe5997b4f5f56e71a330ed7354d5cd2