Don't use wheel when constructing the base mirror
Clients of the mirror may need to fall back on non-wheel files,
and if an upstream project has published wheels, then the mirror
will be missing those in the non-wheel portion. Additionally,
use-wheel given to the wheel command will cause transitive depends
to be skipped in the local wheelhouse.

The only thing that should be making/injecting wheels is the wheel
creation portion of things.

Change-Id: I5e3c4754097249e52f013f6122ff86f01568243c