Pin pip so download cache location works
The new release of pip does not allow you to specify
a dowload cache location - it is deprecated. This means
that the tarball mirror cannot be created because it
relies on the download cache being in the configured
location. This change pins the versin of pip to 1.5.6
so the download cache location still works.

Change-Id: Ib4d058f27de3c37e0c58c50516e6af2b3f8188a5